+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerLicenseServiceImplTest.java
@@ -3,7 +3,6 @@ package de.codescape.jira.plugins.scrumpoker.service;
 import com.atlassian.upm.api.license.PluginLicenseManager;
 import com.atlassian.upm.api.license.entity.LicenseError;
 import com.atlassian.upm.api.license.entity.PluginLicense;
-import com.atlassian.upm.api.util.Option;
 import org.junit.Rule;
 import org.junit.Test;
 import org.mockito.InjectMocks;
@@ -11,6 +10,8 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
+import static com.atlassian.upm.api.util.Option.none;
+import static com.atlassian.upm.api.util.Option.option;
 import static de.codescape.jira.plugins.scrumpoker.service.ScrumPokerLicenseServiceImpl.ERROR_MESSAGE_PREFIX;
 import static de.codescape.jira.plugins.scrumpoker.service.ScrumPokerLicenseServiceImpl.MISSING_LICENSE;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -33,23 +34,23 @@ public class ScrumPokerLicenseServiceImplTest {
 
     @Test
     public void hasValidLicenseReturnsFalseForMissingLicense() {
-        when(pluginLicenseManager.getLicense()).thenReturn(Option.none());
+        when(pluginLicenseManager.getLicense()).thenReturn(none());
         assertThat(scrumPokerLicenseService.hasValidLicense(), is(false));
     }
 
     @Test
     public void hasValidLicenseReturnsFalseForLicenseWithError() {
         PluginLicense pluginLicense = mock(PluginLicense.class);
-        when(pluginLicense.getError()).thenReturn(Option.option(LicenseError.EXPIRED));
-        when(pluginLicenseManager.getLicense()).thenReturn(Option.option(pluginLicense));
+        when(pluginLicense.getError()).thenReturn(option(LicenseError.EXPIRED));
+        when(pluginLicenseManager.getLicense()).thenReturn(option(pluginLicense));
         assertThat(scrumPokerLicenseService.hasValidLicense(), is(false));
     }
 
     @Test
     public void hasValidLicenseReturnsTrueForLicenseWithoutErrors() {
         PluginLicense pluginLicense = mock(PluginLicense.class);
-        when(pluginLicense.getError()).thenReturn(Option.none());
-        when(pluginLicenseManager.getLicense()).thenReturn(Option.option(pluginLicense));
+        when(pluginLicense.getError()).thenReturn(none());
+        when(pluginLicenseManager.getLicense()).thenReturn(option(pluginLicense));
         assertThat(scrumPokerLicenseService.hasValidLicense(), is(true));
     }
 
@@ -57,7 +58,7 @@ public class ScrumPokerLicenseServiceImplTest {
 
     @Test
     public void getLicenseErrorReturnsMissingForMissingLicense() {
-        when(pluginLicenseManager.getLicense()).thenReturn(Option.none());
+        when(pluginLicenseManager.getLicense()).thenReturn(none());
         assertThat(scrumPokerLicenseService.getLicenseError(), is(equalTo(MISSING_LICENSE)));
     }
 
@@ -65,8 +66,8 @@ public class ScrumPokerLicenseServiceImplTest {
     public void getLicenseErrorReturnsErrorForLicenseWithError() {
         LicenseError expectedLicenseError = LicenseError.EXPIRED;
         PluginLicense pluginLicense = mock(PluginLicense.class);
-        when(pluginLicense.getError()).thenReturn(Option.option(expectedLicenseError));
-        when(pluginLicenseManager.getLicense()).thenReturn(Option.option(pluginLicense));
+        when(pluginLicense.getError()).thenReturn(option(expectedLicenseError));
+        when(pluginLicenseManager.getLicense()).thenReturn(option(pluginLicense));
         assertThat(scrumPokerLicenseService.getLicenseError(),
             is(equalTo(ERROR_MESSAGE_PREFIX + expectedLicenseError.name().toLowerCase())));
     }
@@ -74,8 +75,8 @@ public class ScrumPokerLicenseServiceImplTest {
     @Test
     public void getLicenseErrorReturnsNullForLicenseWithoutErrors() {
         PluginLicense pluginLicense = mock(PluginLicense.class);
-        when(pluginLicense.getError()).thenReturn(Option.none());
-        when(pluginLicenseManager.getLicense()).thenReturn(Option.option(pluginLicense));
+        when(pluginLicense.getError()).thenReturn(none());
+        when(pluginLicenseManager.getLicense()).thenReturn(option(pluginLicense));
         assertThat(scrumPokerLicenseService.getLicenseError(), is(nullValue()));
     }