+++ docs/changelog.md
@@ -12,6 +12,7 @@ category: Administration
 ### [Unreleased]
 
 * feat: ensure compatibility with Jira 9.12.0
+* feat: ensure compatibility with Jira 9.12.1 and 9.4.14
 
 ### [23.11.0] - 2023-11-26